♻️ Move TxOptions and DefaultTxOptions to opts
